Test Done

+++ Ankita/OnlineCourierManagementApplication/src/main/java/com/cg/ocma/entities/BankAccount.java
@@ -37,7 +37,7 @@ public class BankAccount {
 	public long getAccountno() {
 		return accountno;
 	}
-	public void setAccountno(int accountno) {
+	public void setAccountno(long accountno) {
 		this.accountno = accountno;
 	}
 	public String getAccountHolderNmae() {
